+++ internal/openrouter/config.go
@@ -3,6 +3,7 @@ package openrouter
 import (
 	"log"
 	"os"
+	"strconv"
 
 	"github.com/joho/godotenv"
 )
@@ -11,6 +12,7 @@ type OpenrouterConfig struct {
 	host   string
 	apiKey string
 	urls   Urls
+	limit  int
 }
 
 type Urls struct {
@@ -25,11 +27,13 @@ type Endpoints struct {
 type EnvVars struct {
 	host   string
 	apiKey string
+	limit  string
 }
 
 var envVars = EnvVars{
 	host:   "OPENROUTER_HOST",
 	apiKey: "OPENROUTER_API_KEY",
+	limit:  "OPENROUTER_LIMIT",
 }
 
 var urls = Urls{
@@ -52,10 +56,24 @@ func getConfig() OpenrouterConfig {
 		log.Fatal("OPENROUTER_API_KEY env variable not provided")
 	}
 
+	limitString := os.Getenv(envVars.limit)
+	if limitString == "" {
+		log.Fatal("OPENROUTER_LIMIT env variable not provided")
+	}
+	if limitString == "-1" {
+		log.Println("WARNING: OPENROUTER_LIMIT is set to -1, which means the limit is infinite")
+	}
+
+	limit, err := strconv.Atoi(limitString)
+	if err != nil {
+		log.Fatal("OPENROUTER_LIMIT env variable could not be converted to int")
+	}
+
 	openrouterConfig := OpenrouterConfig{
 		host:   hostString,
 		apiKey: apiKeyString,
 		urls:   urls,
+		limit:  limit,
 	}
 
 	return openrouterConfig